attendance-tracker/backend/pyproject.toml
[tool.poetry]
name = "attendance-tracker-backend"
version = "0.1.0"
description = "Geofence-based attendance tracking system"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.95.0"
uvicorn = "^0.22.0"
sqlalchemy = "^2.0.0"
pydantic = {extras = ["email"], version = "^1.10.7"}
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
bcrypt = "^4.0.1"
alembic = "^1.10.0"
psycopg2-binary = "^2.9.6"
haversine = "^2.8.0"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.2.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"